feat(backend): document chunking and indexing (#278)

+++ backend/app/repositories/connector.py
@@ -21,16 +21,16 @@ class ConnectorRepository(BaseRepository):
         Connector repository class for handling connector-related database operations.
 
         Args:
-            db_session (Session): Database session
+            db_session (Session): Database session.
         """
         super().__init__(db_session=db_session)
 
     def get_connectors(self) -> Tuple[List[Connector], Optional[APIError]]:
         """
-        Get all connectors
+        Get all connectors.
 
         Returns:
-            Tuple[List[Connector], Optional[APIError]]: List of connector objects and APIError object if any error
+            Tuple[List[Connector], Optional[APIError]]: List of connector objects and APIError object if any error.
         """
         try:
             connectors = self._db_session.query(Connector).all()
@@ -41,18 +41,20 @@ class ConnectorRepository(BaseRepository):
 
     def get_connector(self, connector_id: str) -> Tuple[Optional[Connector], Optional[APIError]]:
         """
-        Get connector by connector_id
+        Get connector by id.
 
         Args:
-            connector_id(str): Connector id
+            connector_id(str): Connector id.
 
         Returns:
-            Tuple[Optional[Connector], Optional[APIError]]: Connector object and APIError object if any error
+            Tuple[Optional[Connector], Optional[APIError]]: Connector object and APIError object if any error.
         """
         try:
             connector = (
                 self._db_session.query(Connector).filter(Connector.id == connector_id).first()
             )
+            if not connector:
+                return None, APIError(kind=ErrorCodesMappingNumber.CONNECTOR_NOT_FOUND.value)
             return connector, None
         except Exception as e:
             logger.error(f"Error getting connector: {e}", exc_info=True)
@@ -60,13 +62,13 @@ class ConnectorRepository(BaseRepository):
 
     def create_connector(self, connector: Connector) -> Optional[APIError]:
         """
-        Create connector
+        Create new connector.
 
         Args:
-            connector(Connector): Connector object
+            connector(Connector): Connector object.
 
         Returns:
-            Optional[APIError]: APIError object if any error
+            Optional[APIError]: APIError object if any error.
         """
         try:
             self._db_session.add(connector)
@@ -77,14 +79,14 @@ class ConnectorRepository(BaseRepository):
 
     def update_connector(self, connector_id: str, connector: Dict[str, Any]) -> Optional[APIError]:
         """
-        Update connector by connector_id
+        Update connector by id.
 
         Args:
-            connector_id(str): Connector id
-            connector(Dict[str, Any]): Connector object
+            connector_id(str): Connector id.
+            connector(Dict[str, Any]): Connector object.
 
         Returns:
-            Optional[APIError]: APIError object if any error
+            Optional[APIError]: APIError object if any error.
         """
         try:
             # Check if connector exists
@@ -103,15 +105,23 @@ class ConnectorRepository(BaseRepository):
 
     def delete_connector(self, connector_id: str) -> Optional[APIError]:
         """
-        Delete connector by connector_id
+        Delete connector by id.
 
         Args:
-            connector_id(str): Connector id
+            connector_id(str): Connector id.
 
         Returns:
-            Optional[APIError]: APIError object if any error
+            Optional[APIError]: APIError object if any error.
         """
         try:
+            # Check if connector exists
+            connector_exists = (
+                self._db_session.query(Connector).filter(Connector.id == connector_id).first()
+            )
+            if not connector_exists:
+                return APIError(kind=ErrorCodesMappingNumber.CONNECTOR_NOT_FOUND.value)
+
+            # Delete connector
             self._db_session.query(Connector).filter(Connector.id == connector_id).delete()
             return None
         except Exception as e:
